requirements.txt
numpy==1.22.4
pandas==1.5.3
streamlit==1.21.0
Pillow==9.5.0

newspaper3k
urllib3
protobuf==3.20.0
jinja2>=3.0

